% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genbinom.R
\name{GenBinomial}
\alias{GenBinomial}
\alias{dgenbinom}
\alias{pgenbinom}
\alias{qgenbinom}
\alias{rgenbinom}
\title{The Generalised Binomial Distribution}
\usage{
dgenbinom(g, pi, theta, size)

pgenbinom(g, pi, theta, size, lower.tail = TRUE)

qgenbinom(p, pi, theta, size, lower.tail = TRUE)

rgenbinom(n, pi, theta, size)
}
\arguments{
\item{g}{vector of quantiles.}

\item{pi}{probability of success of the bernoulli.}

\item{theta}{probability of success on each binomial trial.}

\item{size}{number of trials.}

\item{lower.tail}{logical; if TRUE (default), probabilities are
\eqn{Pr[X \leq x]}{Pr[X \le x]}, otherwise \eqn{Pr[X > x]}{Pr[X > x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\description{
Density, distribution function, quantile function and random
    generation for the generalised binomial distribution with parameters
    \code{pi}, \code{theta} and \code{size}.
}
\details{
The generalised binomial distribution with \code{size = n},
    \code{pi = pi} and \code{theta = theta} has density
    \deqn{Pr(G = g) = \begin{cases} x, g = 0 \\ y, g > 0\end{cases}}
}
